requirement.txt
groq==0.9.0
python-dotenv==1.0.1
langchain==0.2.6
langchain-community==0.0.6
langchain-core==0.2.10
langchain-groq==0.1.2
langchain-openai==0.1.12
langchain-text-splitters==0.2.2
openai==1.35.7
ydantic==2.5.3
pydantic_core==2.14.6